pyproject.toml
[project]
name = "local-llm-research-agent"
version = "0.1.0"
description = "A 100% local smart chat agent for SQL Server data analytics using Ollama and MCP"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
keywords = [
    "llm",
    "ollama",
    "sql-server",
    "mcp",
    "pydantic-ai",
    "chat-agent",
    "data-analytics"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    # Agent Framework
    "pydantic-ai>=0.2.0",
    "pydantic>=2.0.0",

    # Web UI
    "streamlit>=1.30.0",

    # CLI
    "typer>=0.9.0",
    "rich>=13.0.0",

    # Configuration & Environment
    "python-dotenv>=1.0.0",
    "pydantic-settings>=2.0.0",
    "pyyaml>=6.0.0",

    # Logging
    "structlog>=24.0.0",

    # HTTP Client (for Ollama and Foundry Local)
    "httpx>=0.27.0",

    # ===== Phase 2.1 Dependencies =====
    # FastAPI Backend
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-multipart>=0.0.12",
    "aiofiles>=24.1.0",
    "websockets>=14.0",

    # Database (SQLAlchemy for ORM, Alembic for migrations)
    "sqlalchemy>=2.0.0",
    "alembic>=1.14.0",
    "aioodbc>=0.5.0",

    # Redis Vector Store
    "redis>=5.0.0",
    "redisvl>=0.3.0",

    # Document Processing (RAG Pipeline) - Lightweight offline-capable libraries
    "pypdf>=5.0.0",
    "python-docx>=1.1.0",

    # Scheduling
    "apscheduler>=3.10.0",

    # ===== Phase 4.5 Dependencies =====
    # Authentication
    "pyjwt>=2.8.0",
    "passlib[bcrypt]>=1.7.4",
]

[project.optional-dependencies]
foundry = [
    # Microsoft Foundry Local SDK for alternative LLM provider
    "foundry-local-sdk>=0.1.0",
]
postgres = [
    # PostgreSQL async driver for multi-database support
    "asyncpg>=0.29.0",
]
mysql = [
    # MySQL async driver for multi-database support
    "aiomysql>=0.2.0",
]
all-databases = [
    # All database drivers for full multi-database support
    "asyncpg>=0.29.0",
    "aiomysql>=0.2.0",
]
dev = [
    # Testing
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.0.0",
    "locust>=2.0.0",

    # Linting & Formatting
    "ruff>=0.1.0",
    "mypy>=1.0.0",

    # Debugging
    "ipdb>=0.13.0",
]

[project.scripts]
llm-chat = "src.cli.chat:app"

[project.urls]
Homepage = "https://github.com/yourusername/local-llm-research-agent"
Documentation = "https://github.com/yourusername/local-llm-research-agent#readme"
Repository = "https://github.com/yourusername/local-llm-research-agent"
Issues = "https://github.com/yourusername/local-llm-research-agent/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # Pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "SIM", # flake8-simplify
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # do not perform function calls in argument defaults
    "B904",  # raise from in exception handlers (typer.Exit doesn't support it)
]

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["."]
addopts = [
    "-v",
    "--tb=short",
    "-ra",
]
filterwarnings = [
    "ignore::DeprecationWarning",
]
markers = [
    "unit: mark test as unit test (fast, no external dependencies)",
    "integration: mark test as integration test (may use mocked external services)",
    "slow: mark test as slow running",
    "requires_ollama: mark test as requiring a running Ollama instance",
    "requires_foundry: mark test as requiring a running Foundry Local instance",
]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_configs = true
show_error_codes = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = [
    "streamlit.*",
    "typer.*",
    "rich.*",
    "locust.*",
]
ignore_missing_imports = true

[tool.coverage.run]
source = ["src"]
branch = true
omit = [
    "*/tests/*",
    "*/__init__.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]

[tool.bandit]
exclude_dirs = ["tests", ".venv", "venv"]
skips = ["B101"]  # Skip assert warnings in non-test code

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.0.0",
    "locust>=2.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "ipdb>=0.13.0",
    "pre-commit>=3.5.0",
    "bandit[toml]>=1.7.0",
]
